ecommerce_project/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
